Updated routes and added logout functionality
