+++ Apps/AdminWebClient/src/ClientApp/src/App.vue
@@ -3,16 +3,18 @@
     <ToolbarComponent />
 
     <MainNavbar title="HealthGateway Admin" />
+
     <v-content>
       <router-view />
     </v-content>
+
     <Footer />
   </v-app>
 </template>
 
 <script lang="ts">
-import MainNavbar from "@/components/MainNavbar.vue";
-import Footer from "@/components/Footer.vue";
+import MainNavbar from "@/components/core/MainNavbar.vue";
+import Footer from "@/components/core/Footer.vue";
 import ToolbarComponent from "@/components/core/Toolbar.vue";
 
 import { Component, Vue } from "vue-property-decorator";
