+++ Apps/AdminWebClient/src/ClientApp/src/views/Login.vue
@@ -1,21 +1,34 @@
 <template>
-  <v-container> REDIRECTING...</v-container>
+  <v-layout class="fill-height">
+    <LoadingComponent :is-loading="isLoading"></LoadingComponent>
+
+    <v-row justify="center">Redirecting...</v-row>
+  </v-layout>
 </template>
 
 <script lang="ts">
 import { Component, Vue } from "vue-property-decorator";
 import { State, Action, Getter } from "vuex-class";
+import LoadingComponent from "@/components/core/Loading.vue";
 
 const namespace: string = "auth";
 
-@Component
+@Component({
+  components: {
+    LoadingComponent
+  }
+})
 export default class LoginView extends Vue {
   public name = "Dashboard";
   @Action("login", { namespace }) login: any;
   @Getter("isAuthenticated", { namespace }) isAuthenticated: boolean;
+
+  private isLoading: boolean = true;
   private redirectPath: string = "";
   private routeHandler = "";
+
   mounted() {
+    this.isLoading = true;
     if (this.$route.query.redirect && this.$route.query.redirect !== "") {
       this.redirectPath = this.$route.query.redirect;
     } else {
