+++ Apps/AdminWebClient/src/appsettings.Development.json
@@ -42,34 +42,12 @@
     "Immunization": "http://localhost:3001/",
     "Patient": "http://localhost:3002/",
     "Medication": "http://localhost:3003/",
-    "Lab": "http://localhost:3004/"
+    "Lab": "http://localhost:3004/",
+    "JobScheduler": "http://localhost:5005/"
   },
   "ConnectionStrings": {
     "GatewayConnection": "Server=localhost;Port=5432;Database=gateway;User ID=gateway;Password=passw0rd;Integrated Security=true;Pooling=true;"
   },
-  "IdentityProviders": [
-    {
-      "id": "BCSC",
-      "Name": "BC Services Card",
-      "Icon": "address-card",
-      "Hint": "bcsc",
-      "disabled": "false"
-    },
-    {
-      "id": "IDIR",
-      "Name": "BC Government IDIR",
-      "Icon": "user",
-      "Hint": "idir",
-      "disabled": "false"
-    },
-    {
-      "id": "KeyCloak",
-      "Name": "KeyCloak",
-      "Icon": "user-secret",
-      "Hint": "",
-      "disabled": "false"
-    }
-  ],
   "ForwardProxies": {
     "Enabled": "true",
     "BasePath": "",
