+++ Apps/AdminWebClient/src/appsettings.json
@@ -29,24 +29,11 @@
         "IDPHintKey": "kc_idp_hint",
         "IDPHint": "idir"
     },
-    "IdentityProviders": [
-        {
-            "id": "BCSC",
-            "Name": "BC Services Card",
-            "Icon": "address-card",
-            "Hint": "bcsc",
-            "disabled": "false"
-        }
-    ],
     "Admin": {
         "LogLevel": "Debug",
         "TimeOuts": {
             "Idle": "900000",
             "LogoutRedirect": "20000"
-        },
-        "ExternalURLs": {
-            "HealthLinkImmunizationSchedule": "https://www.healthlinkbc.ca/tools-videos/bc-immunization-schedules",
-            "HealthLinkVaccineSearch": "https://www.healthlinkbc.ca/search/"
         }
     },
     "ForwardProxies": {
